+++ packages/wepy/test/fake/com_c.js
@@ -1,0 +1,14 @@
+let assert = require('assert');
+
+let wepy = require('../../lib/wepy.js').default;
+
+module.exports = class Com extends wepy.component {
+
+
+    constructor () {
+        super();
+
+        this.props = ['combprop1', 'combprop2'];
+    }
+
+}