+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.4.5 (waiting)
+## 1.4.4 (2017-02-28)
 * `F` 修复`$invoke`事件`event`参数错误的BUG。[PR #50](https://github.com/wepyjs/wepy/issues/50)
 * `F` 修复`onShareAppMessage`报错的问题。[PR #52](https://github.com/wepyjs/wepy/issues/52)
 * `F` 修复了在没有申明`data`时使用`props`报错的BUG。[PR #54](https://github.com/wepyjs/wepy/issues/54)